Fixed copy bug for windows batch files.
When copy is used on Window 11, the timestampe of the original file is
used instead of a new timestamp. Meaning that all files will have the
same timestampe as when this repo was cloned. The fix was to convert
`copy` commands to `type` commands, and pipe to the output file.